pyproject.toml
[tool.poetry]
name = "modme-clone"
version = "0.1.0"
description = ""
authors = ["Ahror <uzkod136@mail.ru>"]
readme = "README.md"
packages = [{ include = "core" }]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.5"
psycopg2-binary = "^2.9.5"
environ = "^1.0"
django-environ = "^0.9.0"
drf-yasg = { extras = ["validation"], version = "^1.21.4" }
sentry-sdk = "^1.14.0"
requests = "^2.28.2"
pendulum = "^2.1.2"
pillow = "^9.4.0"
djangorestframework-simplejwt = "^5.2.2"
django-filter = "^22.1"
pytest-django = "^4.5.2"
timedelta = "^2020.12.3"
django-rest-auth = "^0.9.5"
django-rest-knox = "^4.2.0"
parler = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
